Removed caching ability from LevelSequence._node_keys(). Benchmarks show a 10% decrease in performance in the worst cases, which does not seem worth a loss of code clarity.
(In particular, I believe that for very large Funcstructs, the time
to handle ints with thousands of digits eclipses time spent computing
parent nodes. For degeneracies where this is not an issue, computation
is fast enough to be irrelevant. Being able to think of these methods
atomically is thus more important).