Function objects are now automatically promoted to the most derived possible types.
For example, if Function is called on input which is invertible but
not endomorphic, it returns a Bijection. If it is endomorphic and
not invertible, it returns an Endofunction. If it is both, it
returns a Permutation.

The same goes for all of Function's subclasses. The only difference
when calling (i.e.) Endofunction directly is that if the Function
is not endomorphic, an error is raised; thus each subclass
enforces that the input Function has the specified behavior.